+++ pylinac/core/image.py
@@ -406,6 +406,7 @@ class BaseImage:
         path : str
             The path to the image.
         """
+        super().__init__()
         self.metrics = []
         self.metric_values = {}
         if isinstance(path, (str, Path)) and not osp.isfile(path):
@@ -1834,7 +1835,7 @@ class LazyZipDicomImageStack(LazyDicomImageStack):
 
     def __init__(
         self,
-        folder: str | Path,
+        folder: str | Path | BinaryIO,
         dtype: np.dtype | None = None,
         min_number: int = 39,
         check_uid: bool = True,
@@ -1847,7 +1848,7 @@ class LazyZipDicomImageStack(LazyDicomImageStack):
         See the documentation for DicomImageStack for parameter descriptions.
         """
         self.dtype = dtype
-        self.zip_archive = Path(folder)
+        self.zip_archive = folder
         self.shadow_images = {}
         with ZipFile(self.zip_archive) as zfile:
             paths = zfile.namelist()
@@ -1877,7 +1878,9 @@ class LazyZipDicomImageStack(LazyDicomImageStack):
         self.create_shadow(paths)
 
     @classmethod
-    def from_zip(cls, zip_path: str | Path, dtype: np.dtype | None = None, **kwargs):
+    def from_zip(
+        cls, zip_path: str | Path | BinaryIO, dtype: np.dtype | None = None, **kwargs
+    ):
         # the zip lazy stack assumes a zip file is passed
         return cls(zip_path, dtype, **kwargs)
 
@@ -2109,7 +2112,7 @@ def tiff_to_dicom(
 
 
 def load_raw_visionrt(
-    path: str | Path, shape: tuple[int, int] = (960, 600), dtype=np.uint32, **kwargs
+    path: str | Path, shape: tuple[int, int] = (600, 960), dtype=np.uint32, **kwargs
 ) -> ArrayImage:
     """Load a .raw file from a VisionRT system.